requirements.txt
asterisk-ami